backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
httpx==0.25.2
python-dotenv==1.0.0
gunicorn==21.2.0
